requirements.txt
flask==1.1.2
pillow==8.0.1
tensorflow==2.3.1
gunicorn==20.0.4